+++ src/main/java/com/universe/service/impl/ProfessorServiceImpl.java
@@ -2,8 +2,10 @@ package com.universe.service.impl;
 
 import com.universe.dto.professor.ProfessorDto;
 import com.universe.entity.ProfessorEntity;
+import com.universe.exception.ConflictException;
 import com.universe.exception.ResourceNotFoundException;
 import com.universe.mapping.ProfessorMapper;
+import com.universe.repository.LessonRepository;
 import com.universe.repository.ProfessorRepository;
 import com.universe.repository.UserRepository;
 import com.universe.service.ProfessorService;
@@ -14,6 +16,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.validation.annotation.Validated;
+
+import static com.universe.repository.LessonRepository.Specs.byProfessorId;
+import static java.lang.String.format;
 
 
 @Service
@@ -27,6 +32,8 @@ public class ProfessorServiceImpl implements ProfessorService {
 
     private final UserRepository userRepository;
 
+    private final LessonRepository lessonRepo;
+
     @Override
     @Transactional(readOnly = true)
     public Page<ProfessorDto> findAll(Pageable pageable) {
@@ -36,33 +43,46 @@ public class ProfessorServiceImpl implements ProfessorService {
 
     @Override
     public ProfessorDto create(ProfessorDto professorDto) {
+        validateUserExistsByEmail(professorDto.getEmail());
         log.info("Creating professor: {}", professorDto);
         var professorEntity = MAPPER.mapBaseAttributes(professorDto);
         var createdProfessor = professorRepo.save(professorEntity);
-
-//        if (!professorDto.getCourses().isEmpty()) {
-//            professorDto.getCourses()
-//                    .forEach(courseDto -> assignProfessorOnCourse(createdProfessor, courseDto.getId()));
-//        }
-
         return MAPPER.mapToDto(createdProfessor);
     }
 
     @Override
     public ProfessorDto update(Long id, ProfessorDto professorDto) {
-        return null;
+        log.info("Updating professor with ID [{}]. Payload: {}", id, professorDto);
+        var professorEntity = findProfessorEntity(id);
+        if (!professorEntity.getEmail().equals(professorDto.getEmail())) {
+            validateUserExistsByEmail(professorDto.getEmail());
+        }
+
+        MAPPER.updateProfessorFromDto(professorDto, professorEntity);
+        var updatedProfessor = professorRepo.save(professorEntity);
+
+        return MAPPER.mapToDto(updatedProfessor);
     }
 
     @Override
     @Transactional(readOnly = true)
     public ProfessorDto findOne(Long professorId) {
         var professorEntity = findProfessorEntity(professorId);
+        var professorLessons = lessonRepo.findAll(byProfessorId(professorId));
+        if (!professorLessons.isEmpty()) {
+            throw new ConflictException("Can't delete professor. Professor has assigned lessons!");
+        }
         return MAPPER.mapToDto(professorEntity);
     }
 
     @Override
     public void delete(Long professorId) {
         var professorEntity = findProfessorEntity(professorId);
+        var lessons = lessonRepo.findAll(byProfessorId(professorId));
+        if (!lessons.isEmpty()) {
+            throw new ConflictException("Can't delete professor. Professor has assigned lessons!");
+        }
+
         professorRepo.delete(professorEntity);
     }
 
@@ -71,5 +91,11 @@ public class ProfessorServiceImpl implements ProfessorService {
                 .orElseThrow(
                         () -> new ResourceNotFoundException(String.format("Professor with id: %s not found!", professorId))
                 );
+    }
+
+    private void validateUserExistsByEmail(String email) {
+        if (userRepository.existsByEmail(email)) {
+            throw new ConflictException(format("User with email: %s already exists!", email));
+        }
     }
 }
